move all_protocols.py to tests
